Patch the init container with compute resources
Fix for #83.
Previously, the compute resources specified in values.yaml did not get passed down to the initContainer specs. When the namespace is guarded by a resource quota, each container needs to specify the resource requirement, otherwise, k8s throws quota errors.

Since the resources definition is a multi-line string block, the fix tries to pass a json formatted string as an argument to the cli.